+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/DefaultMediaSourceFactory.java
@@ -509,23 +509,6 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
   }
 
   /**
-   * Sets whether to enable experimental HAGC (ST 2094-50) metadata playback support. When enabled,
-   * the player will automatically merge HAGC metadata tracks with the associated video track and
-   * deliver the metadata out-of-band to the decoder on API 37+.
-   *
-   * <p>The default value is {@code true}.
-   *
-   * @param enableHagcPlayback Whether experimental HAGC metadata playback is enabled.
-   * @return This factory, for convenience.
-   */
-  @CanIgnoreReturnValue
-  @UnstableApi
-  public DefaultMediaSourceFactory setExperimentalEnableHagcPlayback(boolean enableHagcPlayback) {
-    delegateFactoryLoader.setExperimentalEnableHagcPlayback(enableHagcPlayback);
-    return this;
-  }
-
-  /**
    * Sets whether an experimental setting to delegate end position clipping to a wrapped {@link
    * MediaPeriod} is enabled.
    *
@@ -735,7 +718,6 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
     private SubtitleParser.Factory subtitleParserFactory;
     private @C.VideoCodecFlags int codecsToParseWithinGopSampleDependencies;
     private boolean loadOnlySelectedTracks;
-    private boolean experimentalEnableHagcPlayback;
     @Nullable private CmcdConfiguration.Factory cmcdConfigurationFactory;
     @Nullable private DrmSessionManagerProvider drmSessionManagerProvider;
     @Nullable private LoadErrorHandlingPolicy loadErrorHandlingPolicy;
@@ -749,7 +731,6 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
       mediaSourceFactories = new HashMap<>();
       parseSubtitlesDuringExtraction = true;
       loadOnlySelectedTracks = true;
-      experimentalEnableHagcPlayback = true;
       codecsToParseWithinGopSampleDependencies = C.VIDEO_CODEC_FLAG_H264 | C.VIDEO_CODEC_FLAG_H265;
     }
 
@@ -853,10 +834,6 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
       this.loadOnlySelectedTracks = loadOnlySelectedTracks;
     }
 
-    private void setExperimentalEnableHagcPlayback(boolean experimentalEnableHagcPlayback) {
-      this.experimentalEnableHagcPlayback = experimentalEnableHagcPlayback;
-    }
-
     private void setHeifExtractorFlags(@HeifExtractor.Flags int flags) {
       if (this.extractorsFactory instanceof DefaultExtractorsFactory) {
         ((DefaultExtractorsFactory) this.extractorsFactory).setHeifExtractorFlags(flags);
@@ -930,8 +907,7 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
           mediaSourceFactorySupplier =
               () ->
                   new ProgressiveMediaSource.Factory(dataSourceFactory, extractorsFactory)
-                      .setLoadOnlySelectedTracks(loadOnlySelectedTracks)
-                      .setExperimentalEnableHagcPlayback(experimentalEnableHagcPlayback);
+                      .setLoadOnlySelectedTracks(loadOnlySelectedTracks);
           break;
         default:
           throw new IllegalArgumentException("Unrecognized contentType: " + contentType);